move tool list and core logic out of sse server internals
